+++ torchmetrics/classification/auroc.py
@@ -19,6 +19,7 @@ from torch import Tensor
 from torchmetrics.functional.classification.auroc import _auroc_compute, _auroc_update
 from torchmetrics.metric import Metric
 from torchmetrics.utilities import rank_zero_warn
+from torchmetrics.utilities.data import dim_zero_cat
 from torchmetrics.utilities.imports import _TORCH_LOWER_1_6
 
 
@@ -138,8 +139,8 @@ class AUROC(Metric):
                 )
 
         self.mode = None
-        self.add_state("preds", default=[], dist_reduce_fx=None)
-        self.add_state("target", default=[], dist_reduce_fx=None)
+        self.add_state("preds", default=[], dist_reduce_fx="cat")
+        self.add_state("target", default=[], dist_reduce_fx="cat")
 
         rank_zero_warn(
             'Metric `AUROC` will save all targets and predictions in buffer.'
@@ -170,8 +171,8 @@ class AUROC(Metric):
         """
         Computes AUROC based on inputs passed in to ``update`` previously.
         """
-        preds = torch.cat(self.preds, dim=0)
-        target = torch.cat(self.target, dim=0)
+        preds = dim_zero_cat(self.preds)
+        target = dim_zero_cat(self.target)
         return _auroc_compute(
             preds,
             target,
@@ -181,3 +182,11 @@ class AUROC(Metric):
             self.average,
             self.max_fpr,
         )
+
+    @property
+    def is_differentiable(self):
+        """
+        AUROC metrics is considered as non differentiable so it should have `false`
+        value for `is_differentiable` property
+        """
+        return False