+++ torchmetrics/regression/pearson.py
@@ -19,6 +19,7 @@ from torch import Tensor
 from torchmetrics.functional.regression.pearson import _pearson_corrcoef_compute, _pearson_corrcoef_update
 from torchmetrics.metric import Metric
 from torchmetrics.utilities import rank_zero_warn
+from torchmetrics.utilities.data import dim_zero_cat
 
 
 class PearsonCorrcoef(Metric):
@@ -73,8 +74,8 @@ class PearsonCorrcoef(Metric):
             ' For large datasets this may lead to large memory footprint.'
         )
 
-        self.add_state("preds", default=[], dist_reduce_fx=None)
-        self.add_state("target", default=[], dist_reduce_fx=None)
+        self.add_state("preds", default=[], dist_reduce_fx="cat")
+        self.add_state("target", default=[], dist_reduce_fx="cat")
 
     def update(self, preds: Tensor, target: Tensor):
         """
@@ -92,6 +93,10 @@ class PearsonCorrcoef(Metric):
         """
         Computes pearson correlation coefficient over state.
         """
-        preds = torch.cat(self.preds, dim=0)
-        target = torch.cat(self.target, dim=0)
+        preds = dim_zero_cat(self.preds)
+        target = dim_zero_cat(self.target)
         return _pearson_corrcoef_compute(preds, target)
+
+    @property
+    def is_differentiable(self):
+        return True