+++ torchmetrics/utilities/data.py
@@ -13,7 +13,6 @@
 # limitations under the License.
 from typing import Any, Callable, List, Mapping, Optional, Sequence, Union
 
-import numpy as np
 import torch
 from torch import Tensor, tensor
 
@@ -22,16 +21,17 @@ from torchmetrics.utilities.prints import rank_zero_warn
 METRIC_EPS = 1e-6
 
 
-def dim_zero_cat(x):
+def dim_zero_cat(x: Union[Tensor, List[Tensor]]) -> Tensor:
     x = x if isinstance(x, (list, tuple)) else [x]
+    x = [y.unsqueeze(0) if y.numel() == 1 and y.ndim == 0 else y for y in x]
     return torch.cat(x, dim=0)
 
 
-def dim_zero_sum(x):
+def dim_zero_sum(x: Tensor) -> Tensor:
     return torch.sum(x, dim=0)
 
 
-def dim_zero_mean(x):
+def dim_zero_mean(x: Tensor) -> Tensor:
     return torch.mean(x, dim=0)
 
 
@@ -202,23 +202,22 @@ def apply_to_collection(
     return data
 
 
-def get_group_indexes(indexes: Union[Tensor, np.ndarray]) -> List[Union[Tensor, np.ndarray]]:
+def get_group_indexes(indexes: Tensor) -> List[Tensor]:
     """
-    Given an integer `torch.Tensor` or `np.ndarray` `indexes`, return a `torch.Tensor` or `np.ndarray` of indexes for
+    Given an integer `torch.Tensor` `indexes`, return a `torch.Tensor` of indexes for
     each different value in `indexes`.
 
     Args:
-        indexes: a `torch.Tensor` or `np.ndarray` of integers
+        indexes: a `torch.Tensor`
 
     Return:
-        A list of integer `torch.Tensor`s or `np.ndarray`s
+        A list of integer `torch.Tensor`s
 
     Example:
         >>> indexes = torch.tensor([0, 0, 0, 1, 1, 1, 1])
         >>> get_group_indexes(indexes)
         [tensor([0, 1, 2]), tensor([3, 4, 5, 6])]
     """
-    structure, dtype = (tensor, torch.long) if isinstance(indexes, Tensor) else (np.array, np.int64)
 
     res = dict()
     for i, _id in enumerate(indexes):
@@ -228,4 +227,4 @@ def get_group_indexes(indexes: Union[Tensor, np.ndarray]) -> List[Union[Tensor, 
         else:
             res[_id] = [i]
 
-    return [structure(x, dtype=dtype) for x in res.values()]
+    return [tensor(x, dtype=torch.long) for x in res.values()]